# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://4.233.148.160:8000
    description: Production server
paths:
  /adduser:
    post:
      summary: Add a new user to the database.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's first name.
                  example: Angel
                lastName:
                  type: string
                  description: User's last name.
                  example: Alvarez
                email:
                  type: string
                  description: User's email address.
                  example: angel@example.com
                username:
                  type: string
                  description: Username (required).
                  example: student
                password:
                  type: string
                  description: User's password (required).
                  example: password
                role:
                  type: string
                  description: User's role.
                  enum:
                    - admin
                    - user
                  example: user
      responses:
        '200':
          description: User added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: User's unique identifier.
                    example: 65f756db3fa22d227a4b7c7d
                  name:
                    type: string
                    description: User's first name.
                    example: Angel
                  lastName:
                    type: string
                    description: User's last name.
                    example: Alvarez
                  email:
                    type: string
                    description: User's email address.
                    example: angel@example.com
                  username:
                    type: string
                    description: User's username.
                    example: student
                  role:
                    type: string
                    description: User's role.
                    example: user
                  createdAt:
                    type: string
                    format: date-time
                    description: Date and time when the user was created.
                    example: '2024-03-17T20:47:23.935Z'
                  __v:
                    type: integer
                    description: Internal version key.
                    example: 0
        '400':
          description: Failed to add user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: getaddrinfo EAI_AGAIN mongodb
  /health:
    get:
      summary: Check the health status of the service.
      operationId: checkHealth
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Health status.
                    example: OK
  /login:
    post:
      summary: Log in to the system.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: password
      responses:
        '200':
          description: Login successful. Returns user token, username, and creation date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWY3NTZkYjNmYTIyZDIyN2E0YjdjN2QiLCJpYXQiOjE3MTA3MDg3NDUsImV4cCI6MTcxMDcxMjM0NX0.VMG_5DOyQ4GYlJQRcu1I6ICG1IGzuo2Xuei093ONHxw
                  username:
                    type: string
                    description: Username.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info..
                    example: Invalid credentials
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users:
    get:
      summary: Get all users
      description: Returns a list of all registered users.
      responses:
        '200':
          description: A list of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: The unique ID of the user.
                    name:
                      type: string
                      description: The first name of the user.
                    lastName:
                      type: string
                      description: The last name of the user.
                    email:
                      type: string
                      description: The email address of the user.
                    username:
                      type: string
                      description: The username of the user.
                    role:
                      type: string
                      description: The role of the user (e.g., "admin", "user").
                    createdAt:
                      type: string
                      format: date-time
                      description: The creation date of the user.
                    __v:
                      type: integer
                      description: The version key for the user.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /user:
    get:
      summary: Get a user by ID
      description: Retrieve a user's details by their unique identifier (_id).
      parameters:
        - name: _id
          in: query
          required: true
          description: The unique ID of the user to retrieve.
          schema:
            type: string
            example: 65f756db3fa22d227a4b7c7d
      responses:
        '200':
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: The unique ID of the user.
                    example: 65f756db3fa22d227a4b7c7d
                  name:
                    type: string
                    description: The first name of the user.
                    example: Angel
                  lastName:
                    type: string
                    description: The last name of the user.
                    example: Alvarez
                  email:
                    type: string
                    description: The email address of the user.
                    example: angel@example.com
                  username:
                    type: string
                    description: The username of the user.
                    example: angelalvarez
                  role:
                    type: string
                    description: The role of the user (e.g., "admin", "user").
                    example: user
                  createdAt:
                    type: string
                    format: date-time
                    description: The creation date of the user.
                    example: '2024-03-17T20:47:23.935Z'
                  __v:
                    type: integer
                    description: The version key for the user.
                    example: 0
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message indicating the user was not found.
                    example: User not found.
        '400':
          description: Invalid ID supplied.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message indicating an invalid ID.
                    example: Invalid ID format.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /questions:
    get:
      summary: Retrieve all questions
      description: Fetches a list of all the questions stored in the system.
      responses:
        '200':
          description: List of questions successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: Unique identifier for the question
                    type:
                      type: string
                      description: Type of the question (e.g., "multiple-choice")
                    name:
                      type: string
                      description: Name or title of the question (optional)
                    path:
                      type: string
                      description: Path or URL associated with the question
                    hint1:
                      type: string
                      description: First hint for the question (optional)
                    hint2:
                      type: string
                      description: Second hint for the question (optional)
                    right:
                      type: string
                      description: Correct answer for the question
                    wrong1:
                      type: string
                      description: First incorrect answer
                    wrong2:
                      type: string
                      description: Second incorrect answer
                    wrong3:
                      type: string
                      description: Third incorrect answer
                    createdAt:
                      type: string
                      format: date-time
                      description: Date when the question was created
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /stats:
    get:
      summary: Get all game statistics
      description: Returns a list of all game statistics.
      responses:
        '200':
          description: A list of game statistics.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: The unique ID of the statistic record.
                    userId:
                      type: string
                      description: The ID of the user associated with the record.
                    gameId:
                      type: string
                      description: The ID of the game associated with the record.
                    questionId:
                      type: string
                      description: The ID of the question associated with the record.
                    right:
                      type: boolean
                      description: Whether the answer was correct.
                    time:
                      type: integer
                      description: The time taken to answer in seconds.
                    points:
                      type: integer
                      description: The points scored for the answer.
                    createdAt:
                      type: string
                      format: date-time
                      description: The creation date of the statistic record.
                    __v:
                      type: integer
                      description: The version key for the statistic record.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /ranking:
    get:
      summary: Get the top ranking list
      description: Retrieve the top ranking list with game results, including points and time for each user.
      responses:
        '200':
          description: A list of top-ranking users and their game results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  topRanking:
                    type: array
                    description: A list of top-ranking items.
                    items:
                      type: object
                      properties:
                        gameId:
                          type: string
                          description: The unique ID of the game.
                          example: a4acc100-52d6-46ab-8ce2-b0acccf93c03
                        userId:
                          type: string
                          description: The unique ID of the user who played the game.
                          example: 6767e65426af19f147062821
                        totalPoints:
                          type: integer
                          description: The total points scored by the user in the game.
                          example: 2600
                        totalTime:
                          type: integer
                          description: The total time (in seconds) spent by the user on the game.
                          example: 49
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error